+++ src/main/groovy/ru/vyarus/gradle/plugin/quality/QualityExtension.groovy
@@ -23,7 +23,7 @@ class QualityExtension {
 
     String checkstyleVersion = '8.40'
     String pmdVersion = '6.31.0'
-    String spotbugsVersion = '4.2.0'
+    String spotbugsVersion = '4.2.1'
     String codenarcVersion = '2.0.0'
 
     /**